🗺️ Atlas: Splitting Temporal Index Blob
🕸️ Tangle: The structural problem is that `src/index/temporal.rs` was a 4200+ line "Blob", mixing the core temporal index logic with 2800+ lines of tests, making it unwieldy.
📐 Blueprint: Extracted `src/index/temporal.rs` into a module directory `src/index/temporal/`, placing the core code in `mod.rs` and the tests in `tests.rs`.
🧱 Stability: Improved maintainability and reduced file bloat without changing any functionality.
🔬 Verification: Builds successfully, all tests pass, strictly checked formatting and documentation.
